Page Fixes and buttons added
now only gives item achievements when getting the correct item, now resets the items on generating a new bingo page, and added buttons to directly take player to the advancements page and back. also added a reset button for players who join/rejoin later if advancements are not synced properly and criterion is incorrectly completed.
back button exists but is currently invisible in our subclass